% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{materialize}
\alias{materialize}
\alias{size}
\alias{import}
\alias{cached_table}
\alias{db_table}
\alias{query_table}
\alias{pquery_table}
\alias{import_all}
\alias{table_references}
\title{Inspect and import/cache database tables as tibbles.}
\usage{
materialize(reference, ...)

size(reference, ...)

import(reference, ignore_cache = FALSE, global = TRUE,
  overwrite = TRUE)

cached_table(name)

db_table(conn, name)

query_table(conn, query, name)

pquery_table(conn, query, name, query_parameters)

import_all(...)

table_references(ref_type, ref_parameters = l(), ...)
}
\arguments{
\item{reference}{a lazy table reference.}

\item{name}{a string containing the name of the lazy table reference. For
`db_table` and `cached_table`, this has to match the name of the table in 
the database and/or the name of the cache file.}

\item{conn}{a database connection created with `DBI::dbConnect`. Can be 
set to `NULL` if the user is sure there is a cached version
on disk and the reference is being used with `import`.}

\item{query}{a string containing a (possibly parametric) SQL query.}

\item{query_parameters}{a list of named parameters (e.g. 
`list(par1='value1', par2='value2')` to be substituted into the query.}

\item{ref_type}{a lazy reference type. Either `db_table`, `cached_table`, or,
less commonly, `parametric_table` or `cached_table`.}

\item{ref_parameters}{a list of named parameters to be passed to `ref_type`
with each call.}
}
\description{
Utility functions for reading remote database tables (using SQL) into 
tibbles and caching them into disk for future use.
}
\details{
\describe{
   \item{`db_table()`}{lazy reference to a database table which can 
   be read into memory as a tibble and cached into disk.}

\item{`query_table()`}{lazy reference to a query result which can
   be read into memory as a tibble and cached into disk.}
   
   \item{`pquery_table()`}{same as `query_table()`, but taking a
   parametric query and a list of named query parameters.}
   
   \item{`cached_table()`}{lazy reference to a table which has been
   previously cached. This exists mainly so that cached tables can 
   be read without setting up a database connection, which is required
   by all other reference types.}
   
   \item{`materialize()`}{materializes a lazy reference into an actual
   tibble by accessing the remote database. This function never reads
   or writes to the local cache.}
   
   \item{`size()`}{returns the size, in megabytes, of a lazy table reference 
   without materializing it. This may be useful for probing the size of a 
   database table before deciding to download it.}
   
   \item{`import()`}{materializes a lazy reference into an actual tibble
   by reading from the local cache, or calling `materialize()` and then 
   caching results if nothing is cached. The resulting tible will be 
   automatically bound to a variable with the table's `name` in the global
   or parent environment.}
   
   \item{`import_all()`}{utility function. Equivalent to looping through a
   set of lazy references and calling import in each.}
   
   \item{`cached_tables()`}{utility function for creating several cached
   table references from their names alone.}
}
}
