% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{is_weekend}
\alias{is_weekend}
\title{Tells if a given date falls on a weekend or not.}
\usage{
is_weekend(date)
}
\arguments{
\item{a}{date or datetime object.}
}
\description{
Tells if a given date falls on a weekend or not.
}
