% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{paste_vec}
\alias{paste_vec}
\title{Concatenate a vector of strings}
\usage{
paste_vec(.dots, sep = "")
}
\description{
`paste_vec` works the same as \link{paste}, but instead of taking arguments 
directly (i.e. using `...`) it takes them as a vector.
}
\examples{

data_vec <- c('hello', 'world')

# Same as do.call(paste, c(data_vec, list(sep = ' ')))
paste_vec(data_vec, sep = ' ')

}
