% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concurrency.R
\name{realCores}
\alias{realCores}
\title{Simple wrapper to \link{parallel::detectCores} which yields the number of
physical cores in Linux as well.}
\usage{
realCores()
}
\description{
Simple wrapper to \link{parallel::detectCores} which yields the number of
physical cores in Linux as well.
}
