% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{weekday_find_closest}
\alias{weekday_find_closest}
\title{Given a reference date object and a weekday as a string or numeric index, finds
the closest instance of that weekday that happened before the reference date.}
\usage{
weekday_find_closest(weekday, reference = NULL)
}
\arguments{
\item{weekday}{a day of the week, as a string or an index. These are 
locale-specific. Use \link{weekday_labels} to figure out the valid 
labels for your system.}

\item{reference}{a reference date from where to look for the required 
weekday.}
}
\description{
For instance, if the reference date is a Friday and the weekday is "Saturday",
the function will return last Saturday.
}
