% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{weekday_label}
\alias{weekday_label}
\title{Convert datetime to weekday label}
\usage{
weekday_label(datetime)
}
\description{
Returns locale-specific day-of-the-week labels for a date/datetime vector.
}
\examples{

datetime <- c('2020-04-23', '2020-04-24')
weekday_label(datetime)

}
\seealso{
weekday_labels
}
