% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{diff_na}
\alias{diff_na}
\title{Lagged Differences}
\usage{
diff_na(x, ...)
}
\description{
A version of \code{\link{diff}} which returns a vector
of the same length as the original (beginning with a NA),
and does not propagates NAs. \code{diff_na} takes the same
parameters as \code{\link{diff}}.
}
\seealso{
\code{\link{diff}} for valid parameters to \code{diff_na}.
}
