% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcs.R
\name{gcs_data}
\alias{gcs_data}
\alias{gcs_table}
\alias{materialize.gcs_table}
\alias{gcs_auth}
\title{Access datasets from Google Cloud Storage}
\usage{
gcs_data(bucket, path, loader)

gcs_table(bucket, path, loader)

\method{materialize}{gcs_table}(reference)

gcs_auth()
}
\description{
Helper functions for loading datasets from Google Cloud Storage. In case of
tabular data, provides functions which can be used in concert with 
\link{import}.
}
