% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{roll_stat}
\alias{roll_stat}
\title{Computes rolling statistics}
\usage{
roll_stat(x, n, stat)
}
\arguments{
\item{x}{a vector}

\item{n}{a window size}
}
\description{
Computes a rolling statistic `stat` over windows of size `n` taken from
vector `x`.
}
\examples{

  x <- c(1, 0, 1, 1, 0, 1, 0, 0, 0)
  # returns 2 2 2 2 1 1 0
  roll_stat(x, n = 3, sum)

}
